# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang GoFiber swagger auto generate step by step by swaggo",
        "title": "GoFiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Endpoint untuk login user dan mengembalikan data user beserta token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "Login User",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succeed to POST data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.LoginRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Endpoint untuk register user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "description": "Register User",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil login, mengembalikan data user",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CityResp": {
            "description": "Data kota yang terkait dengan user",
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "Bandung"
                },
                "province_id": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "dto.CreateUser": {
            "type": "object",
            "required": [
                "email",
                "id_kota",
                "id_provinsi",
                "jenis_kelamin",
                "kata_sandi",
                "name",
                "no_telp",
                "pekerjaan",
                "tanggal_lahir",
                "tentang"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id_kota": {
                    "type": "string"
                },
                "id_provinsi": {
                    "type": "string"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "kata_sandi": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "pekerjaan": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "tentang": {
                    "type": "string"
                }
            }
        },
...
        "dto.Login": {
            "type": "object",
            "required": [
                "kata_sandi",
                "no_telp"
            ],
            "properties": {
                "kata_sandi": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                }
            }
        },
        "dto.LoginRes": {
            "description": "Data yang dikembalikan setelah user berhasil login",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "L2DQK@example.com"
                },
                "id_kota": {
                    "$ref": "#/definitions/dto.CityResp"
                },
                "id_provinsi": {
                    "$ref": "#/definitions/dto.ProvinceResp"
                },
                "nama": {
                    "type": "string",
                    "example": "John Doe"
                },
                "no_telp": {
                    "type": "string",
                    "example": "1234567890"
                },
                "pekerjaan": {
                    "type": "string",
                    "example": "Software Engineer"
                },
                "tanggal_lahir": {
                    "type": "string",
                    "example": "1990-01-01"
                },
                "tentang": {
                    "type": "string",
                    "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "dto.ProvinceResp": {
            "description": "Data provinsi yang terkait dengan user",
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "Jawa Barat"
                }
            }
        },
...
        "helper.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {}
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}